+++ staging/src/k8s.io/client-go/plugin/pkg/client/auth/openstack/BUILD
@@ -3,25 +3,15 @@ package(default_visibility = ["//visibility:public"])
 load(
     "@io_bazel_rules_go//go:def.bzl",
     "go_library",
-    "go_test",
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["openstack_test.go"],
-    embed = [":go_default_library"],
 )
 
 go_library(
     name = "go_default_library",
-    srcs = ["openstack.go"],
+    srcs = ["openstack_stub.go"],
     importmap = "k8s.io/kubernetes/vendor/k8s.io/client-go/plugin/pkg/client/auth/openstack",
     importpath = "k8s.io/client-go/plugin/pkg/client/auth/openstack",
     deps = [
-        "//staging/src/k8s.io/apimachinery/pkg/util/net:go_default_library",
         "//staging/src/k8s.io/client-go/rest:go_default_library",
-        "//vendor/github.com/gophercloud/gophercloud:go_default_library",
-        "//vendor/github.com/gophercloud/gophercloud/openstack:go_default_library",
         "//vendor/k8s.io/klog:go_default_library",
     ],
 )
